catalog: Restructure durable catalog files
This commit shuffles around the files within the mz_catalog crate. This
is in preparation for a generic trait that will be implemented using
the stash. The changes are purely code movement and changes to
visibility modifiers.

Works towards resolving #20953
