+++ src/catalog/src/objects.rs
@@ -7,241 +7,68 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
+use mz_audit_log::{VersionedEvent, VersionedStorageUsage};
+use mz_compute_client::logging::LogVariant;
+use mz_controller::clusters::ReplicaLogging;
+use mz_controller_types::{ClusterId, ReplicaId};
 use mz_ore::cast::CastFrom;
 use mz_proto::{IntoRustIfSome, ProtoType};
+use mz_repr::adt::mz_acl_item::{AclMode, MzAclItem};
+use mz_repr::role_id::RoleId;
+use mz_repr::GlobalId;
+use mz_sql::catalog::{CatalogItemType, ObjectType, RoleAttributes, RoleMembership};
+use mz_sql::names::{CommentObjectId, DatabaseId, QualifiedItemName, SchemaId};
 use mz_stash::objects::{proto, RustType, TryFromProtoError};
+use proptest_derive::Arbitrary;
+use std::time::Duration;
 
-use crate::{
-    AuditLogKey, ClusterConfig, ClusterIntrospectionSourceIndexKey,
-    ClusterIntrospectionSourceIndexValue, ClusterKey, ClusterReplicaKey, ClusterReplicaValue,
-    ClusterValue, ClusterVariant, ClusterVariantManaged, CommentKey, CommentValue, ConfigKey,
-    ConfigValue, DatabaseKey, DatabaseValue, DefaultPrivilegesKey, DefaultPrivilegesValue,
-    GidMappingKey, GidMappingValue, IdAllocKey, IdAllocValue, ItemKey, ItemValue, ReplicaConfig,
-    ReplicaLocation, RoleKey, RoleValue, SchemaKey, SchemaValue, ServerConfigurationKey,
-    ServerConfigurationValue, SettingKey, SettingValue, StorageUsageKey, SystemPrivilegesKey,
-    SystemPrivilegesValue, TimestampKey, TimestampValue,
-};
+// Structs used to pass information to outside modules.
 
-impl RustType<proto::ConfigKey> for ConfigKey {
-    fn into_proto(&self) -> proto::ConfigKey {
-        proto::ConfigKey {
-            key: self.key.to_string(),
-        }
-    }
-
-    fn from_proto(proto: proto::ConfigKey) -> Result<Self, TryFromProtoError> {
-        Ok(ConfigKey { key: proto.key })
-    }
+#[derive(Debug, Clone)]
+pub struct Database {
+    pub id: DatabaseId,
+    pub name: String,
+    pub owner_id: RoleId,
+    pub privileges: Vec<MzAclItem>,
 }
 
-impl RustType<proto::ConfigValue> for ConfigValue {
-    fn into_proto(&self) -> proto::ConfigValue {
-        proto::ConfigValue { value: self.value }
-    }
-
-    fn from_proto(proto: proto::ConfigValue) -> Result<Self, TryFromProtoError> {
-        Ok(ConfigValue { value: proto.value })
-    }
+#[derive(Debug, Clone)]
+pub struct Schema {
+    pub id: SchemaId,
+    pub name: String,
+    pub database_id: Option<DatabaseId>,
+    pub owner_id: RoleId,
+    pub privileges: Vec<MzAclItem>,
 }
 
-impl RustType<proto::SettingKey> for SettingKey {
-    fn into_proto(&self) -> proto::SettingKey {
-        proto::SettingKey {
-            name: self.name.to_string(),
-        }
-    }
-
-    fn from_proto(proto: proto::SettingKey) -> Result<Self, TryFromProtoError> {
-        Ok(SettingKey { name: proto.name })
-    }
+#[derive(Debug, Clone)]
+pub struct Role {
+    pub id: RoleId,
+    pub name: String,
+    pub attributes: RoleAttributes,
+    pub membership: RoleMembership,
 }
 
-impl RustType<proto::SettingValue> for SettingValue {
-    fn into_proto(&self) -> proto::SettingValue {
-        proto::SettingValue {
-            value: self.value.to_string(),
-        }
-    }
-
-    fn from_proto(proto: proto::SettingValue) -> Result<Self, TryFromProtoError> {
-        Ok(SettingValue { value: proto.value })
-    }
+#[derive(Clone, Debug)]
+pub struct BuiltinRole {
+    pub id: RoleId,
+    pub name: &'static str,
+    pub attributes: RoleAttributes,
 }
 
-impl RustType<proto::IdAllocKey> for IdAllocKey {
-    fn into_proto(&self) -> proto::IdAllocKey {
-        proto::IdAllocKey {
-            name: self.name.to_string(),
-        }
-    }
-
-    fn from_proto(proto: proto::IdAllocKey) -> Result<Self, TryFromProtoError> {
-        Ok(IdAllocKey { name: proto.name })
-    }
+#[derive(Debug, Clone)]
+pub struct Cluster {
+    pub id: ClusterId,
+    pub name: String,
+    pub linked_object_id: Option<GlobalId>,
+    pub owner_id: RoleId,
+    pub privileges: Vec<MzAclItem>,
+    pub config: ClusterConfig,
 }
 
-impl RustType<proto::IdAllocValue> for IdAllocValue {
-    fn into_proto(&self) -> proto::IdAllocValue {
-        proto::IdAllocValue {
-            next_id: self.next_id,
-        }
-    }
-
-    fn from_proto(proto: proto::IdAllocValue) -> Result<Self, TryFromProtoError> {
-        Ok(IdAllocValue {
-            next_id: proto.next_id,
-        })
-    }
-}
-
-impl RustType<proto::GidMappingKey> for GidMappingKey {
-    fn into_proto(&self) -> proto::GidMappingKey {
-        proto::GidMappingKey {
-            schema_name: self.schema_name.to_string(),
-            object_type: self.object_type.into_proto().into(),
-            object_name: self.object_name.to_string(),
-        }
-    }
-
-    fn from_proto(proto: proto::GidMappingKey) -> Result<Self, TryFromProtoError> {
-        let object_type = proto::CatalogItemType::from_i32(proto.object_type)
-            .ok_or_else(|| TryFromProtoError::unknown_enum_variant("CatalogItemType"))?;
-        Ok(GidMappingKey {
-            schema_name: proto.schema_name,
-            object_type: object_type.into_rust()?,
-            object_name: proto.object_name,
-        })
-    }
-}
-
-impl RustType<proto::GidMappingValue> for GidMappingValue {
-    fn into_proto(&self) -> proto::GidMappingValue {
-        proto::GidMappingValue {
-            id: self.id,
-            fingerprint: self.fingerprint.to_string(),
-        }
-    }
-
-    fn from_proto(proto: proto::GidMappingValue) -> Result<Self, TryFromProtoError> {
-        Ok(GidMappingValue {
-            id: proto.id,
-            fingerprint: proto.fingerprint,
-        })
-    }
-}
-
-impl RustType<proto::ClusterKey> for ClusterKey {
-    fn into_proto(&self) -> proto::ClusterKey {
-        proto::ClusterKey {
-            id: Some(self.id.into_proto()),
-        }
-    }
-
-    fn from_proto(proto: proto::ClusterKey) -> Result<Self, TryFromProtoError> {
-        Ok(ClusterKey {
-            id: proto.id.into_rust_if_some("ClusterKey::id")?,
-        })
-    }
-}
-
-impl RustType<proto::ClusterValue> for ClusterValue {
-    fn into_proto(&self) -> proto::ClusterValue {
-        proto::ClusterValue {
-            name: self.name.to_string(),
-            config: Some(self.config.into_proto()),
-            linked_object_id: self.linked_object_id.into_proto(),
-            owner_id: Some(self.owner_id.into_proto()),
-            privileges: self.privileges.into_proto(),
-        }
-    }
-
-    fn from_proto(proto: proto::ClusterValue) -> Result<Self, TryFromProtoError> {
-        Ok(ClusterValue {
-            name: proto.name,
-            config: proto.config.unwrap_or_default().into_rust()?,
-            linked_object_id: proto.linked_object_id.into_rust()?,
-            owner_id: proto.owner_id.into_rust_if_some("ClusterValue::owner_id")?,
-            privileges: proto.privileges.into_rust()?,
-        })
-    }
-}
-
-impl RustType<proto::ClusterIntrospectionSourceIndexKey> for ClusterIntrospectionSourceIndexKey {
-    fn into_proto(&self) -> proto::ClusterIntrospectionSourceIndexKey {
-        proto::ClusterIntrospectionSourceIndexKey {
-            cluster_id: Some(self.cluster_id.into_proto()),
-            name: self.name.to_string(),
-        }
-    }
-
-    fn from_proto(
-        proto: proto::ClusterIntrospectionSourceIndexKey,
-    ) -> Result<Self, TryFromProtoError> {
-        Ok(ClusterIntrospectionSourceIndexKey {
-            cluster_id: proto
-                .cluster_id
-                .into_rust_if_some("ClusterIntrospectionSourceIndexKey::cluster_id")?,
-            name: proto.name,
-        })
-    }
-}
-
-impl RustType<proto::ClusterIntrospectionSourceIndexValue>
-    for ClusterIntrospectionSourceIndexValue
-{
-    fn into_proto(&self) -> proto::ClusterIntrospectionSourceIndexValue {
-        proto::ClusterIntrospectionSourceIndexValue {
-            index_id: self.index_id,
-        }
-    }
-
-    fn from_proto(
-        proto: proto::ClusterIntrospectionSourceIndexValue,
-    ) -> Result<Self, TryFromProtoError> {
-        Ok(ClusterIntrospectionSourceIndexValue {
-            index_id: proto.index_id,
-        })
-    }
-}
-
-impl RustType<proto::ClusterReplicaKey> for ClusterReplicaKey {
-    fn into_proto(&self) -> proto::ClusterReplicaKey {
-        proto::ClusterReplicaKey {
-            id: Some(self.id.into_proto()),
-        }
-    }
-
-    fn from_proto(proto: proto::ClusterReplicaKey) -> Result<Self, TryFromProtoError> {
-        Ok(ClusterReplicaKey {
-            id: proto.id.into_rust_if_some("ClusterReplicaKey::id")?,
-        })
-    }
-}
-
-impl RustType<proto::ClusterReplicaValue> for ClusterReplicaValue {
-    fn into_proto(&self) -> proto::ClusterReplicaValue {
-        proto::ClusterReplicaValue {
-            cluster_id: Some(self.cluster_id.into_proto()),
-            name: self.name.to_string(),
-            config: Some(self.config.into_proto()),
-            owner_id: Some(self.owner_id.into_proto()),
-        }
-    }
-
-    fn from_proto(proto: proto::ClusterReplicaValue) -> Result<Self, TryFromProtoError> {
-        Ok(ClusterReplicaValue {
-            cluster_id: proto
-                .cluster_id
-                .into_rust_if_some("ClusterReplicaValue::cluster_id")?,
-            name: proto.name,
-            config: proto
-                .config
-                .into_rust_if_some("ClusterReplicaValue::config")?,
-            owner_id: proto
-                .owner_id
-                .into_rust_if_some("ClusterReplicaValue::owner_id")?,
-        })
-    }
+#[derive(Clone, Debug, PartialOrd, PartialEq, Eq, Ord)]
+pub struct ClusterConfig {
+    pub variant: ClusterVariant,
 }
 
 impl RustType<proto::ClusterConfig> for ClusterConfig {
@@ -256,6 +83,12 @@ impl RustType<proto::ClusterConfig> for ClusterConfig {
             variant: proto.variant.into_rust_if_some("ClusterConfig::variant")?,
         })
     }
+}
+
+#[derive(Clone, Debug, PartialOrd, PartialEq, Eq, Ord)]
+pub enum ClusterVariant {
+    Managed(ClusterVariantManaged),
+    Unmanaged,
 }
 
 impl RustType<proto::cluster_config::Variant> for ClusterVariant {
@@ -302,6 +135,51 @@ impl RustType<proto::cluster_config::Variant> for ClusterVariant {
     }
 }
 
+#[derive(Clone, Debug, PartialOrd, PartialEq, Eq, Ord)]
+pub struct ClusterVariantManaged {
+    pub size: String,
+    pub availability_zones: Vec<String>,
+    pub logging: ReplicaLogging,
+    pub idle_arrangement_merge_effort: Option<u32>,
+    pub replication_factor: u32,
+    pub disk: bool,
+}
+
+#[derive(Clone, Debug)]
+pub struct BuiltinCluster {
+    pub name: &'static str,
+    pub privileges: &'static [MzAclItem],
+}
+
+#[derive(Debug, Clone)]
+pub struct ClusterReplica {
+    pub cluster_id: ClusterId,
+    pub replica_id: ReplicaId,
+    pub name: String,
+    pub config: ReplicaConfig,
+    pub owner_id: RoleId,
+}
+
+// The on-disk replica configuration does not match the in-memory replica configuration, so we need
+// separate structs. As of writing this comment, it is mainly due to the fact that we don't persist
+// the replica allocation.
+#[derive(Clone, Debug, PartialOrd, PartialEq, Eq, Ord)]
+pub struct ReplicaConfig {
+    pub location: ReplicaLocation,
+    pub logging: ReplicaLogging,
+    pub idle_arrangement_merge_effort: Option<u32>,
+}
+
+impl From<mz_controller::clusters::ReplicaConfig> for ReplicaConfig {
+    fn from(config: mz_controller::clusters::ReplicaConfig) -> Self {
+        Self {
+            location: config.location.into(),
+            logging: config.compute.logging,
+            idle_arrangement_merge_effort: config.compute.idle_arrangement_merge_effort,
+        }
+    }
+}
+
 impl RustType<proto::ReplicaConfig> for ReplicaConfig {
     fn into_proto(&self) -> proto::ReplicaConfig {
         proto::ReplicaConfig {
@@ -321,6 +199,65 @@ impl RustType<proto::ReplicaConfig> for ReplicaConfig {
             logging: proto.logging.into_rust_if_some("ReplicaConfig::logging")?,
             idle_arrangement_merge_effort: proto.idle_arrangement_merge_effort.map(|e| e.effort),
         })
+    }
+}
+
+#[derive(Debug, Clone, PartialOrd, PartialEq, Eq, Ord)]
+pub enum ReplicaLocation {
+    Unmanaged {
+        storagectl_addrs: Vec<String>,
+        storage_addrs: Vec<String>,
+        computectl_addrs: Vec<String>,
+        compute_addrs: Vec<String>,
+        workers: usize,
+    },
+    Managed {
+        size: String,
+        /// `Some(az)` if the AZ was specified by the user and must be respected;
+        availability_zone: Option<String>,
+        disk: bool,
+    },
+}
+
+impl From<mz_controller::clusters::ReplicaLocation> for ReplicaLocation {
+    fn from(loc: mz_controller::clusters::ReplicaLocation) -> Self {
+        match loc {
+            mz_controller::clusters::ReplicaLocation::Unmanaged(
+                mz_controller::clusters::UnmanagedReplicaLocation {
+                    storagectl_addrs,
+                    storage_addrs,
+                    computectl_addrs,
+                    compute_addrs,
+                    workers,
+                },
+            ) => Self::Unmanaged {
+                storagectl_addrs,
+                storage_addrs,
+                computectl_addrs,
+                compute_addrs,
+                workers,
+            },
+            mz_controller::clusters::ReplicaLocation::Managed(
+                mz_controller::clusters::ManagedReplicaLocation {
+                    allocation: _,
+                    size,
+                    availability_zones,
+                    disk,
+                },
+            ) => ReplicaLocation::Managed {
+                size,
+                availability_zone:
+                    if let mz_controller::clusters::ManagedReplicaAvailabilityZones::FromReplica(
+                        Some(az),
+                    ) = availability_zones
+                    {
+                        Some(az)
+                    } else {
+                        None
+                    },
+                disk,
+            },
+        }
     }
 }
 
@@ -374,6 +311,328 @@ impl RustType<proto::replica_config::Location> for ReplicaLocation {
     }
 }
 
+#[derive(Clone, Debug)]
+pub struct ComputeReplicaLogging {
+    pub log_logging: bool,
+    pub interval: Option<Duration>,
+}
+
+#[derive(Clone, Debug)]
+pub struct BuiltinClusterReplica {
+    pub name: &'static str,
+    pub cluster_name: &'static str,
+}
+
+#[derive(Debug, Clone)]
+pub struct Item {
+    pub id: GlobalId,
+    pub name: QualifiedItemName,
+    pub create_sql: String,
+    pub owner_id: RoleId,
+    pub privileges: Vec<MzAclItem>,
+}
+
+/// Functions can share the same name as any other catalog item type
+/// within a given schema.
+/// For example, a function can have the same name as a type, e.g.
+/// 'date'.
+/// As such, system objects are keyed in the catalog storage by the
+/// tuple (schema_name, object_type, object_name), which is guaranteed
+/// to be unique.
+#[derive(Debug, Clone)]
+pub struct SystemObjectMapping {
+    pub schema_name: String,
+    pub object_type: CatalogItemType,
+    pub object_name: String,
+    pub id: GlobalId,
+    pub fingerprint: String,
+}
+
+#[derive(Clone, Debug)]
+pub struct BuiltinLog {
+    pub variant: LogVariant,
+    pub name: &'static str,
+    pub schema: &'static str,
+}
+
+// Structs used internally to represent on disk-state.
+
+#[derive(Clone, PartialOrd, PartialEq, Eq, Ord, Hash)]
+pub struct SettingKey {
+    pub(crate) name: String,
+}
+
+impl RustType<proto::SettingKey> for SettingKey {
+    fn into_proto(&self) -> proto::SettingKey {
+        proto::SettingKey {
+            name: self.name.to_string(),
+        }
+    }
+
+    fn from_proto(proto: proto::SettingKey) -> Result<Self, TryFromProtoError> {
+        Ok(SettingKey { name: proto.name })
+    }
+}
+
+#[derive(Clone, PartialOrd, PartialEq, Eq, Ord)]
+pub struct SettingValue {
+    pub(crate) value: String,
+}
+
+impl RustType<proto::SettingValue> for SettingValue {
+    fn into_proto(&self) -> proto::SettingValue {
+        proto::SettingValue {
+            value: self.value.to_string(),
+        }
+    }
+
+    fn from_proto(proto: proto::SettingValue) -> Result<Self, TryFromProtoError> {
+        Ok(SettingValue { value: proto.value })
+    }
+}
+
+#[derive(Clone, PartialOrd, PartialEq, Eq, Ord, Hash)]
+pub struct IdAllocKey {
+    pub(crate) name: String,
+}
+
+impl RustType<proto::IdAllocKey> for IdAllocKey {
+    fn into_proto(&self) -> proto::IdAllocKey {
+        proto::IdAllocKey {
+            name: self.name.to_string(),
+        }
+    }
+
+    fn from_proto(proto: proto::IdAllocKey) -> Result<Self, TryFromProtoError> {
+        Ok(IdAllocKey { name: proto.name })
+    }
+}
+
+#[derive(Clone, PartialOrd, PartialEq, Eq, Ord)]
+pub struct IdAllocValue {
+    pub(crate) next_id: u64,
+}
+
+impl RustType<proto::IdAllocValue> for IdAllocValue {
+    fn into_proto(&self) -> proto::IdAllocValue {
+        proto::IdAllocValue {
+            next_id: self.next_id,
+        }
+    }
+
+    fn from_proto(proto: proto::IdAllocValue) -> Result<Self, TryFromProtoError> {
+        Ok(IdAllocValue {
+            next_id: proto.next_id,
+        })
+    }
+}
+
+#[derive(Clone, PartialOrd, PartialEq, Eq, Ord, Hash)]
+pub struct GidMappingKey {
+    pub(crate) schema_name: String,
+    pub(crate) object_type: CatalogItemType,
+    pub(crate) object_name: String,
+}
+
+impl RustType<proto::GidMappingKey> for GidMappingKey {
+    fn into_proto(&self) -> proto::GidMappingKey {
+        proto::GidMappingKey {
+            schema_name: self.schema_name.to_string(),
+            object_type: self.object_type.into_proto().into(),
+            object_name: self.object_name.to_string(),
+        }
+    }
+
+    fn from_proto(proto: proto::GidMappingKey) -> Result<Self, TryFromProtoError> {
+        let object_type = proto::CatalogItemType::from_i32(proto.object_type)
+            .ok_or_else(|| TryFromProtoError::unknown_enum_variant("CatalogItemType"))?;
+        Ok(GidMappingKey {
+            schema_name: proto.schema_name,
+            object_type: object_type.into_rust()?,
+            object_name: proto.object_name,
+        })
+    }
+}
+
+#[derive(Clone, PartialOrd, PartialEq, Eq, Ord)]
+pub struct GidMappingValue {
+    pub(crate) id: u64,
+    pub(crate) fingerprint: String,
+}
+
+impl RustType<proto::GidMappingValue> for GidMappingValue {
+    fn into_proto(&self) -> proto::GidMappingValue {
+        proto::GidMappingValue {
+            id: self.id,
+            fingerprint: self.fingerprint.to_string(),
+        }
+    }
+
+    fn from_proto(proto: proto::GidMappingValue) -> Result<Self, TryFromProtoError> {
+        Ok(GidMappingValue {
+            id: proto.id,
+            fingerprint: proto.fingerprint,
+        })
+    }
+}
+
+#[derive(Clone, PartialOrd, PartialEq, Eq, Ord, Hash)]
+pub struct ClusterKey {
+    pub(crate) id: ClusterId,
+}
+
+impl RustType<proto::ClusterKey> for ClusterKey {
+    fn into_proto(&self) -> proto::ClusterKey {
+        proto::ClusterKey {
+            id: Some(self.id.into_proto()),
+        }
+    }
+
+    fn from_proto(proto: proto::ClusterKey) -> Result<Self, TryFromProtoError> {
+        Ok(ClusterKey {
+            id: proto.id.into_rust_if_some("ClusterKey::id")?,
+        })
+    }
+}
+
+#[derive(Clone, PartialOrd, PartialEq, Eq, Ord)]
+pub struct ClusterValue {
+    pub(crate) name: String,
+    pub(crate) linked_object_id: Option<GlobalId>,
+    pub(crate) owner_id: RoleId,
+    pub(crate) privileges: Vec<MzAclItem>,
+    pub(crate) config: ClusterConfig,
+}
+
+impl RustType<proto::ClusterValue> for ClusterValue {
+    fn into_proto(&self) -> proto::ClusterValue {
+        proto::ClusterValue {
+            name: self.name.to_string(),
+            config: Some(self.config.into_proto()),
+            linked_object_id: self.linked_object_id.into_proto(),
+            owner_id: Some(self.owner_id.into_proto()),
+            privileges: self.privileges.into_proto(),
+        }
+    }
+
+    fn from_proto(proto: proto::ClusterValue) -> Result<Self, TryFromProtoError> {
+        Ok(ClusterValue {
+            name: proto.name,
+            config: proto.config.unwrap_or_default().into_rust()?,
+            linked_object_id: proto.linked_object_id.into_rust()?,
+            owner_id: proto.owner_id.into_rust_if_some("ClusterValue::owner_id")?,
+            privileges: proto.privileges.into_rust()?,
+        })
+    }
+}
+
+#[derive(Clone, PartialOrd, PartialEq, Eq, Ord, Hash)]
+pub struct ClusterIntrospectionSourceIndexKey {
+    pub(crate) cluster_id: ClusterId,
+    pub(crate) name: String,
+}
+
+impl RustType<proto::ClusterIntrospectionSourceIndexKey> for ClusterIntrospectionSourceIndexKey {
+    fn into_proto(&self) -> proto::ClusterIntrospectionSourceIndexKey {
+        proto::ClusterIntrospectionSourceIndexKey {
+            cluster_id: Some(self.cluster_id.into_proto()),
+            name: self.name.to_string(),
+        }
+    }
+
+    fn from_proto(
+        proto: proto::ClusterIntrospectionSourceIndexKey,
+    ) -> Result<Self, TryFromProtoError> {
+        Ok(ClusterIntrospectionSourceIndexKey {
+            cluster_id: proto
+                .cluster_id
+                .into_rust_if_some("ClusterIntrospectionSourceIndexKey::cluster_id")?,
+            name: proto.name,
+        })
+    }
+}
+
+#[derive(Clone, PartialOrd, PartialEq, Eq, Ord)]
+pub struct ClusterIntrospectionSourceIndexValue {
+    pub(crate) index_id: u64,
+}
+
+impl RustType<proto::ClusterIntrospectionSourceIndexValue>
+    for ClusterIntrospectionSourceIndexValue
+{
+    fn into_proto(&self) -> proto::ClusterIntrospectionSourceIndexValue {
+        proto::ClusterIntrospectionSourceIndexValue {
+            index_id: self.index_id,
+        }
+    }
+
+    fn from_proto(
+        proto: proto::ClusterIntrospectionSourceIndexValue,
+    ) -> Result<Self, TryFromProtoError> {
+        Ok(ClusterIntrospectionSourceIndexValue {
+            index_id: proto.index_id,
+        })
+    }
+}
+
+#[derive(Clone, PartialOrd, PartialEq, Eq, Ord, Hash)]
+pub struct ClusterReplicaKey {
+    pub(crate) id: ReplicaId,
+}
+
+impl RustType<proto::ClusterReplicaKey> for ClusterReplicaKey {
+    fn into_proto(&self) -> proto::ClusterReplicaKey {
+        proto::ClusterReplicaKey {
+            id: Some(self.id.into_proto()),
+        }
+    }
+
+    fn from_proto(proto: proto::ClusterReplicaKey) -> Result<Self, TryFromProtoError> {
+        Ok(ClusterReplicaKey {
+            id: proto.id.into_rust_if_some("ClusterReplicaKey::id")?,
+        })
+    }
+}
+
+#[derive(Clone, PartialOrd, PartialEq, Eq, Ord)]
+pub struct ClusterReplicaValue {
+    pub(crate) cluster_id: ClusterId,
+    pub(crate) name: String,
+    pub(crate) config: ReplicaConfig,
+    pub(crate) owner_id: RoleId,
+}
+
+impl RustType<proto::ClusterReplicaValue> for ClusterReplicaValue {
+    fn into_proto(&self) -> proto::ClusterReplicaValue {
+        proto::ClusterReplicaValue {
+            cluster_id: Some(self.cluster_id.into_proto()),
+            name: self.name.to_string(),
+            config: Some(self.config.into_proto()),
+            owner_id: Some(self.owner_id.into_proto()),
+        }
+    }
+
+    fn from_proto(proto: proto::ClusterReplicaValue) -> Result<Self, TryFromProtoError> {
+        Ok(ClusterReplicaValue {
+            cluster_id: proto
+                .cluster_id
+                .into_rust_if_some("ClusterReplicaValue::cluster_id")?,
+            name: proto.name,
+            config: proto
+                .config
+                .into_rust_if_some("ClusterReplicaValue::config")?,
+            owner_id: proto
+                .owner_id
+                .into_rust_if_some("ClusterReplicaValue::owner_id")?,
+        })
+    }
+}
+
+#[derive(Clone, Copy, Debug, PartialOrd, PartialEq, Eq, Ord, Hash, Arbitrary)]
+pub struct DatabaseKey {
+    pub(crate) id: DatabaseId,
+}
+
 impl RustType<proto::DatabaseKey> for DatabaseKey {
     fn into_proto(&self) -> proto::DatabaseKey {
         proto::DatabaseKey {
@@ -386,6 +645,13 @@ impl RustType<proto::DatabaseKey> for DatabaseKey {
             id: proto.id.into_rust_if_some("DatabaseKey::value")?,
         })
     }
+}
+
+#[derive(Clone, Debug, PartialOrd, PartialEq, Eq, Ord, Arbitrary)]
+pub struct DatabaseValue {
+    pub(crate) name: String,
+    pub(crate) owner_id: RoleId,
+    pub(crate) privileges: Vec<MzAclItem>,
 }
 
 impl RustType<proto::DatabaseValue> for DatabaseValue {
@@ -408,6 +674,11 @@ impl RustType<proto::DatabaseValue> for DatabaseValue {
     }
 }
 
+#[derive(Clone, Copy, Debug, PartialOrd, PartialEq, Eq, Ord, Hash, Arbitrary)]
+pub struct SchemaKey {
+    pub(crate) id: SchemaId,
+}
+
 impl RustType<proto::SchemaKey> for SchemaKey {
     fn into_proto(&self) -> proto::SchemaKey {
         proto::SchemaKey {
@@ -420,6 +691,14 @@ impl RustType<proto::SchemaKey> for SchemaKey {
             id: proto.id.into_rust_if_some("SchemaKey::id")?,
         })
     }
+}
+
+#[derive(Clone, Debug, PartialOrd, PartialEq, Eq, Ord, Arbitrary)]
+pub struct SchemaValue {
+    pub(crate) database_id: Option<DatabaseId>,
+    pub(crate) name: String,
+    pub(crate) owner_id: RoleId,
+    pub(crate) privileges: Vec<MzAclItem>,
 }
 
 impl RustType<proto::SchemaValue> for SchemaValue {
@@ -444,6 +723,11 @@ impl RustType<proto::SchemaValue> for SchemaValue {
     }
 }
 
+#[derive(Clone, PartialOrd, PartialEq, Eq, Ord, Hash, Debug, Arbitrary)]
+pub struct ItemKey {
+    pub(crate) gid: GlobalId,
+}
+
 impl RustType<proto::ItemKey> for ItemKey {
     fn into_proto(&self) -> proto::ItemKey {
         proto::ItemKey {
@@ -456,6 +740,15 @@ impl RustType<proto::ItemKey> for ItemKey {
             gid: proto.gid.into_rust_if_some("ItemKey::gid")?,
         })
     }
+}
+
+#[derive(Clone, Debug, PartialOrd, PartialEq, Eq, Ord, Arbitrary)]
+pub struct ItemValue {
+    pub(crate) schema_id: SchemaId,
+    pub(crate) name: String,
+    pub(crate) create_sql: String,
+    pub(crate) owner_id: RoleId,
+    pub(crate) privileges: Vec<MzAclItem>,
 }
 
 impl RustType<proto::ItemValue> for ItemValue {
@@ -493,6 +786,12 @@ impl RustType<proto::ItemValue> for ItemValue {
     }
 }
 
+#[derive(Clone, Debug, PartialOrd, PartialEq, Eq, Ord)]
+pub struct CommentKey {
+    pub(crate) object_id: CommentObjectId,
+    pub(crate) sub_component: Option<usize>,
+}
+
 impl RustType<proto::CommentKey> for CommentKey {
     fn into_proto(&self) -> proto::CommentKey {
         let sub_component = match &self.sub_component {
@@ -521,6 +820,11 @@ impl RustType<proto::CommentKey> for CommentKey {
     }
 }
 
+#[derive(Clone, Debug, PartialOrd, PartialEq, Eq, Ord, Arbitrary)]
+pub struct CommentValue {
+    pub(crate) comment: String,
+}
+
 impl RustType<proto::CommentValue> for CommentValue {
     fn into_proto(&self) -> proto::CommentValue {
         proto::CommentValue {
@@ -535,6 +839,11 @@ impl RustType<proto::CommentValue> for CommentValue {
     }
 }
 
+#[derive(Clone, PartialOrd, PartialEq, Eq, Ord, Hash, Debug)]
+pub struct RoleKey {
+    pub(crate) id: RoleId,
+}
+
 impl RustType<proto::RoleKey> for RoleKey {
     fn into_proto(&self) -> proto::RoleKey {
         proto::RoleKey {
@@ -546,6 +855,23 @@ impl RustType<proto::RoleKey> for RoleKey {
         Ok(RoleKey {
             id: proto.id.into_rust_if_some("RoleKey::id")?,
         })
+    }
+}
+
+#[derive(Clone, PartialOrd, PartialEq, Eq, Ord, Debug)]
+pub struct RoleValue {
+    pub(crate) name: String,
+    pub(crate) attributes: RoleAttributes,
+    pub(crate) membership: RoleMembership,
+}
+
+impl From<Role> for RoleValue {
+    fn from(role: Role) -> Self {
+        RoleValue {
+            name: role.name,
+            attributes: role.attributes,
+            membership: role.membership,
+        }
     }
 }
 
@@ -571,54 +897,60 @@ impl RustType<proto::RoleValue> for RoleValue {
     }
 }
 
-impl RustType<proto::TimestampKey> for TimestampKey {
-    fn into_proto(&self) -> proto::TimestampKey {
-        proto::TimestampKey {
-            id: self.id.clone(),
+#[derive(Debug, Clone, PartialOrd, PartialEq, Eq, Ord)]
+pub struct ConfigKey {
+    pub(crate) key: String,
+}
+
+impl RustType<proto::ConfigKey> for ConfigKey {
+    fn into_proto(&self) -> proto::ConfigKey {
+        proto::ConfigKey {
+            key: self.key.to_string(),
         }
     }
 
-    fn from_proto(proto: proto::TimestampKey) -> Result<Self, TryFromProtoError> {
-        Ok(TimestampKey { id: proto.id })
+    fn from_proto(proto: proto::ConfigKey) -> Result<Self, TryFromProtoError> {
+        Ok(ConfigKey { key: proto.key })
     }
 }
 
-impl RustType<proto::TimestampValue> for TimestampValue {
-    fn into_proto(&self) -> proto::TimestampValue {
-        proto::TimestampValue {
-            ts: Some(self.ts.into_proto()),
+#[derive(Clone, PartialOrd, PartialEq, Eq, Ord, Hash)]
+pub struct ConfigValue {
+    pub(crate) value: u64,
+}
+
+impl RustType<proto::ConfigValue> for ConfigValue {
+    fn into_proto(&self) -> proto::ConfigValue {
+        proto::ConfigValue { value: self.value }
+    }
+
+    fn from_proto(proto: proto::ConfigValue) -> Result<Self, TryFromProtoError> {
+        Ok(ConfigValue { value: proto.value })
+    }
+}
+
+#[derive(Clone, PartialOrd, PartialEq, Eq, Ord, Hash)]
+pub struct AuditLogKey {
+    pub(crate) event: VersionedEvent,
+}
+
+impl RustType<proto::AuditLogKey> for AuditLogKey {
+    fn into_proto(&self) -> proto::AuditLogKey {
+        proto::AuditLogKey {
+            event: Some(self.event.into_proto()),
         }
     }
 
-    fn from_proto(proto: proto::TimestampValue) -> Result<Self, TryFromProtoError> {
-        Ok(TimestampValue {
-            ts: proto.ts.into_rust_if_some("TimestampValue::ts")?,
+    fn from_proto(proto: proto::AuditLogKey) -> Result<Self, TryFromProtoError> {
+        Ok(AuditLogKey {
+            event: proto.event.into_rust_if_some("AuditLogKey::event")?,
         })
     }
 }
 
-impl RustType<proto::ServerConfigurationKey> for ServerConfigurationKey {
-    fn into_proto(&self) -> proto::ServerConfigurationKey {
-        proto::ServerConfigurationKey {
-            name: self.name.clone(),
-        }
-    }
-
-    fn from_proto(proto: proto::ServerConfigurationKey) -> Result<Self, TryFromProtoError> {
-        Ok(ServerConfigurationKey { name: proto.name })
-    }
-}
-
-impl RustType<proto::ServerConfigurationValue> for ServerConfigurationValue {
-    fn into_proto(&self) -> proto::ServerConfigurationValue {
-        proto::ServerConfigurationValue {
-            value: self.value.clone(),
-        }
-    }
-
-    fn from_proto(proto: proto::ServerConfigurationValue) -> Result<Self, TryFromProtoError> {
-        Ok(ServerConfigurationValue { value: proto.value })
-    }
+#[derive(Clone, PartialOrd, PartialEq, Eq, Ord, Hash)]
+pub struct StorageUsageKey {
+    pub(crate) metric: VersionedStorageUsage,
 }
 
 impl RustType<proto::StorageUsageKey> for StorageUsageKey {
@@ -635,18 +967,83 @@ impl RustType<proto::StorageUsageKey> for StorageUsageKey {
     }
 }
 
-impl RustType<proto::AuditLogKey> for AuditLogKey {
-    fn into_proto(&self) -> proto::AuditLogKey {
-        proto::AuditLogKey {
-            event: Some(self.event.into_proto()),
+#[derive(Clone, PartialOrd, PartialEq, Eq, Ord, Hash)]
+pub struct TimestampKey {
+    pub(crate) id: String,
+}
+
+impl RustType<proto::TimestampKey> for TimestampKey {
+    fn into_proto(&self) -> proto::TimestampKey {
+        proto::TimestampKey {
+            id: self.id.clone(),
         }
     }
 
-    fn from_proto(proto: proto::AuditLogKey) -> Result<Self, TryFromProtoError> {
-        Ok(AuditLogKey {
-            event: proto.event.into_rust_if_some("AuditLogKey::event")?,
+    fn from_proto(proto: proto::TimestampKey) -> Result<Self, TryFromProtoError> {
+        Ok(TimestampKey { id: proto.id })
+    }
+}
+
+#[derive(Clone, PartialOrd, PartialEq, Eq, Ord)]
+pub struct TimestampValue {
+    pub(crate) ts: mz_repr::Timestamp,
+}
+
+impl RustType<proto::TimestampValue> for TimestampValue {
+    fn into_proto(&self) -> proto::TimestampValue {
+        proto::TimestampValue {
+            ts: Some(self.ts.into_proto()),
+        }
+    }
+
+    fn from_proto(proto: proto::TimestampValue) -> Result<Self, TryFromProtoError> {
+        Ok(TimestampValue {
+            ts: proto.ts.into_rust_if_some("TimestampValue::ts")?,
         })
     }
+}
+
+#[derive(Clone, PartialOrd, PartialEq, Eq, Ord, Hash)]
+pub struct ServerConfigurationKey {
+    pub(crate) name: String,
+}
+
+impl RustType<proto::ServerConfigurationKey> for ServerConfigurationKey {
+    fn into_proto(&self) -> proto::ServerConfigurationKey {
+        proto::ServerConfigurationKey {
+            name: self.name.clone(),
+        }
+    }
+
+    fn from_proto(proto: proto::ServerConfigurationKey) -> Result<Self, TryFromProtoError> {
+        Ok(ServerConfigurationKey { name: proto.name })
+    }
+}
+
+#[derive(Clone, PartialOrd, PartialEq, Eq, Ord)]
+pub struct ServerConfigurationValue {
+    pub(crate) value: String,
+}
+
+impl RustType<proto::ServerConfigurationValue> for ServerConfigurationValue {
+    fn into_proto(&self) -> proto::ServerConfigurationValue {
+        proto::ServerConfigurationValue {
+            value: self.value.clone(),
+        }
+    }
+
+    fn from_proto(proto: proto::ServerConfigurationValue) -> Result<Self, TryFromProtoError> {
+        Ok(ServerConfigurationValue { value: proto.value })
+    }
+}
+
+#[derive(Clone, PartialOrd, PartialEq, Eq, Ord, Hash)]
+pub struct DefaultPrivilegesKey {
+    pub(crate) role_id: RoleId,
+    pub(crate) database_id: Option<DatabaseId>,
+    pub(crate) schema_id: Option<SchemaId>,
+    pub(crate) object_type: ObjectType,
+    pub(crate) grantee: RoleId,
 }
 
 impl RustType<proto::DefaultPrivilegesKey> for DefaultPrivilegesKey {
@@ -677,6 +1074,11 @@ impl RustType<proto::DefaultPrivilegesKey> for DefaultPrivilegesKey {
     }
 }
 
+#[derive(Clone, PartialOrd, PartialEq, Eq, Ord, Hash)]
+pub struct DefaultPrivilegesValue {
+    pub(crate) privileges: AclMode,
+}
+
 impl RustType<proto::DefaultPrivilegesValue> for DefaultPrivilegesValue {
     fn into_proto(&self) -> proto::DefaultPrivilegesValue {
         proto::DefaultPrivilegesValue {
@@ -691,6 +1093,12 @@ impl RustType<proto::DefaultPrivilegesValue> for DefaultPrivilegesValue {
                 .into_rust_if_some("DefaultPrivilegesValue::privileges")?,
         })
     }
+}
+
+#[derive(Clone, PartialOrd, PartialEq, Eq, Ord, Hash)]
+pub struct SystemPrivilegesKey {
+    pub(crate) grantee: RoleId,
+    pub(crate) grantor: RoleId,
 }
 
 impl RustType<proto::SystemPrivilegesKey> for SystemPrivilegesKey {
@@ -713,6 +1121,11 @@ impl RustType<proto::SystemPrivilegesKey> for SystemPrivilegesKey {
     }
 }
 
+#[derive(Clone, PartialOrd, PartialEq, Eq, Ord, Hash)]
+pub struct SystemPrivilegesValue {
+    pub(crate) acl_mode: AclMode,
+}
+
 impl RustType<proto::SystemPrivilegesValue> for SystemPrivilegesValue {
     fn into_proto(&self) -> proto::SystemPrivilegesValue {
         proto::SystemPrivilegesValue {
@@ -726,5 +1139,69 @@ impl RustType<proto::SystemPrivilegesValue> for SystemPrivilegesValue {
                 .acl_mode
                 .into_rust_if_some("SystemPrivilegesKey::acl_mode")?,
         })
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use mz_proto::{ProtoType, RustType};
+    use proptest::prelude::*;
+
+    use super::{DatabaseKey, DatabaseValue, ItemKey, ItemValue, SchemaKey, SchemaValue};
+
+    proptest! {
+        #[mz_ore::test]
+        #[cfg_attr(miri, ignore)] // slow
+        fn proptest_database_key_roundtrip(key: DatabaseKey) {
+            let proto = key.into_proto();
+            let round = proto.into_rust().expect("to roundtrip");
+
+            prop_assert_eq!(key, round);
+        }
+
+        #[mz_ore::test]
+        #[cfg_attr(miri, ignore)] // slow
+        fn proptest_database_value_roundtrip(value: DatabaseValue) {
+            let proto = value.into_proto();
+            let round = proto.into_rust().expect("to roundtrip");
+
+            prop_assert_eq!(value, round);
+        }
+
+        #[mz_ore::test]
+        #[cfg_attr(miri, ignore)] // slow
+        fn proptest_schema_key_roundtrip(key: SchemaKey) {
+            let proto = key.into_proto();
+            let round = proto.into_rust().expect("to roundtrip");
+
+            prop_assert_eq!(key, round);
+        }
+
+        #[mz_ore::test]
+        #[cfg_attr(miri, ignore)] // slow
+        fn proptest_schema_value_roundtrip(value: SchemaValue) {
+            let proto = value.into_proto();
+            let round = proto.into_rust().expect("to roundtrip");
+
+            prop_assert_eq!(value, round);
+        }
+
+        #[mz_ore::test]
+        #[cfg_attr(miri, ignore)] // slow
+        fn proptest_item_key_roundtrip(key: ItemKey) {
+            let proto = key.into_proto();
+            let round = proto.into_rust().expect("to roundtrip");
+
+            prop_assert_eq!(key, round);
+        }
+
+        #[mz_ore::test]
+        #[cfg_attr(miri, ignore)] // slow
+        fn proptest_item_value_roundtrip(value: ItemValue) {
+            let proto = value.into_proto();
+            let round = proto.into_rust().expect("to roundtrip");
+
+            prop_assert_eq!(value, round);
+        }
     }
 }
